+++ spring-rewrite-commons-launcher/README.adoc
@@ -95,9 +95,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.rewrite.parser.ProjectScanner;
 import org.springframework.rewrite.parser.RewriteProjectParser;
 import org.springframework.rewrite.parser.RewriteProjectParsingResult;
-import org.springframework.rewrite.project.resource.ProjectResourceSet;
-import org.springframework.rewrite.project.resource.ProjectResourceSetFactory;
-import org.springframework.rewrite.project.resource.ProjectResourceSetSerializer;
+import org.springframework.rewrite.resource.ProjectResourceSet;
+import org.springframework.rewrite.resource.ProjectResourceSetFactory;
+import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
 import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
 
 import java.nio.file.Path;
