+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/boot/autoconfigure/RewriteLauncherConfiguration.java
@@ -20,8 +20,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.rewrite.execution.RewriteRecipeLauncher;
 import org.springframework.rewrite.parser.RewriteProjectParser;
-import org.springframework.rewrite.project.resource.ProjectResourceSetFactory;
-import org.springframework.rewrite.project.resource.ProjectResourceSetSerializer;
+import org.springframework.rewrite.resource.ProjectResourceSetFactory;
+import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
 import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
 
 /**
