+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/resource/BaseProjectResource.java
@@ -13,31 +13,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.rewrite.project.resource.finder;
+package org.springframework.rewrite.resource;
 
-import org.springframework.rewrite.project.resource.ProjectResourceSet;
+public abstract class BaseProjectResource implements InternalProjectResource {
 
-import java.util.List;
-import java.util.stream.Collectors;
+	protected boolean isChanged = false;
 
-public class GenericTypeListFinder<T> implements ProjectResourceFinder<List<T>> {
+	private boolean isDeleted = false;
 
-	private final Class<T> type;
-
-	public GenericTypeListFinder(Class<T> type) {
-		this.type = type;
+	public boolean hasChanges() {
+		return isChanged;
 	}
 
-	public Class<T> getType() {
-		return type;
+	public void resetHasChanges() {
+		isChanged = false;
 	}
 
-	@Override
-	public List<T> apply(ProjectResourceSet projectResourceSet) {
-		return projectResourceSet.stream()
-			.filter(pr -> type.isAssignableFrom(pr.getClass()))
-			.map(type::cast)
-			.collect(Collectors.toList());
+	public void markAsChanged() {
+		isChanged = true;
+	}
+
+	public void delete() {
+		isDeleted = true;
+		markAsChanged();
+	}
+
+	public boolean isDeleted() {
+		return isDeleted;
 	}
 
 }
