+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/resource/ProjectResourceSerializer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.rewrite.project.resource;
+package org.springframework.rewrite.resource;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
