+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/resource/RewriteSourceFileHolder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.rewrite.project.resource;
+package org.springframework.rewrite.resource;
 
 import org.openrewrite.SourceFile;
 import org.openrewrite.internal.lang.Nullable;
