+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/resource/finder/GenericTypeListFinder.java
@@ -13,12 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.rewrite.project.resource.finder;
+package org.springframework.rewrite.resource.finder;
 
-public class ResourceFilterException extends RuntimeException {
+import org.springframework.rewrite.resource.ProjectResourceSet;
 
-	public ResourceFilterException(String s) {
-		super(s);
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class GenericTypeListFinder<T> implements ProjectResourceFinder<List<T>> {
+
+	private final Class<T> type;
+
+	public GenericTypeListFinder(Class<T> type) {
+		this.type = type;
+	}
+
+	public Class<T> getType() {
+		return type;
+	}
+
+	@Override
+	public List<T> apply(ProjectResourceSet projectResourceSet) {
+		return projectResourceSet.stream()
+			.filter(pr -> type.isAssignableFrom(pr.getClass()))
+			.map(type::cast)
+			.collect(Collectors.toList());
 	}
 
 }
