+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/resource/finder/PathPatternMatchingProjectResourceFinder.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.rewrite.project.resource.finder;
+package org.springframework.rewrite.resource.finder;
 
-import org.springframework.rewrite.project.resource.ProjectResource;
-import org.springframework.rewrite.project.resource.ProjectResourceSet;
+import org.springframework.rewrite.resource.ProjectResource;
+import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.utils.OsAgnosticPathMatcher;
 import org.springframework.util.PathMatcher;
